requirement.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
blis==0.4.1
boto3==1.14.25
botocore==1.17.25
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cloudpickle==0.6.1
coloredlogs==10.0
constantly==15.1.0
cycler==0.10.0
cymem==2.0.3
Django==3.0.8
djangorestframework==3.11.0
docutils==0.15.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
future==0.17.1
gevent==1.5.0
greenlet==0.4.16
humanfriendly==8.2
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
jmespath==0.10.0
jsonschema==2.6.0
kiwisolver==1.2.0
klein==17.10.0
matplotlib==2.2.5
murmurhash==1.0.2
numpy==1.19.1
packaging==18.0
pandas==1.0.5
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
PyHamcrest==2.0.2
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
rasa-nlu==0.15.1
requests==2.24.0
ruamel.yaml==0.15.100
s3transfer==0.3.3
scikit-learn==0.20.4
scipy==1.5.1
simplejson==3.17.2
six==1.15.0
sklearn-crfsuite==0.3.6
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
tabulate==0.8.7
thinc==7.4.1
tqdm==4.48.0
Twisted==20.3.0
typing==3.7.4.3
urllib3==1.25.9
wasabi==0.7.1
Werkzeug==1.0.1
zipp==3.1.0
zope.interface==5.1.0
